requirements.txt
notion-client>=2.0.0
Pillow>=10.0.0
PyYAML>=6.0
gpiozero>=2.0;platform_machine=='armv7l' or platform_machine=='aarch64'
spidev>=3.6;platform_machine=='armv7l' or platform_machine=='aarch64'
